requirements.txt
google-cloud-dlp==3.3.1
google-cloud-storage==1.43.0